Connection to db is working, Added management of list of Pois for admins
