+++ src/main/java/org/solent/spring/map/controller/MapPointRestController.java
@@ -11,24 +11,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.solent.spring.map.controller;
 
-import org.solent.spring.map.model.MapPoint;
-import org.solent.spring.map.repository.MapPointRepository;
-import org.solent.spring.map.user.model.dto.User;
-import org.solent.spring.map.user.model.dto.UserRole;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
-import javax.servlet.http.HttpSession;
 import java.util.Optional;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.solent.spring.map.model.MapPoint;
+import org.solent.spring.map.repository.MapPointRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
 @RestController
 @RequestMapping("/")
@@ -37,21 +35,7 @@ public class MapPointRestController {
     final static Logger LOG = LogManager.getLogger(MapPointRestController.class);
 
     @Autowired
-    @Qualifier("serviceEntityManagerFactory")
-    private LocalContainerEntityManagerFactoryBean entityManagerFactory;
     private MapPointRepository mapPointRepository;
-
-
-    private User getSessionUser(HttpSession session) {
-        User sessionUser = (User) session.getAttribute("sessionUser");
-        if (sessionUser == null) {
-            sessionUser = new User();
-            sessionUser.setUsername("anonymous");
-            sessionUser.setUserRole(UserRole.ANONYMOUS);
-            session.setAttribute("sessionUser", sessionUser);
-        }
-        return sessionUser;
-    }
 
     @Operation(summary = "Get a list of map points")
     @RequestMapping("/get")
@@ -63,6 +47,6 @@ public class MapPointRestController {
     @RequestMapping("/get/{id}")
     public MapPoint getById(@Parameter(description = "id if point to be retreived") @PathVariable(value = "id") long id) {
         Optional<MapPoint> mpo = mapPointRepository.findById(id);
-        return mpo.orElse(null);
+        return (mpo.isEmpty()) ? null : mpo.get();
     }
 }
