+++ src/main/java/org/solent/spring/map/repository/MapPointRepository.java
@@ -16,6 +16,10 @@ package org.solent.spring.map.repository;
 import org.solent.spring.map.model.MapPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
 
 @Repository
 public interface MapPointRepository extends JpaRepository<MapPoint,Long>{
