+++ src/main/resources/application.properties
@@ -4,11 +4,12 @@ spring.mvc.view.suffix=.jsp
 springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/swagger-ui.html
 
+
 #spring.thymeleaf.cache=false
 # DB Settings
 spring.datasource.url= jdbc:mysql://devops.mysql.database.azure.com/db
-spring.datasource.username= hamza
-spring.datasource.password= admin123_
+spring.datasource.username=hamza
+spring.datasource.password=admin123_
 spring.datasource.driver-class-name= com.mysql.jdbc.Driver
 
 spring.datasource.dbcp.test-while-idle=true
@@ -16,7 +17,7 @@ spring.datasource.dbcp.validation-query = SELECT 1
 
 spring.jpa.show-sql = true
 ## What to do with db on init (create, create-drop, update)
-spring.jpa.hibernate.ddl-auto= create-drop
+spring.jpa.hibernate.ddl-auto= update
