+++ src/main/resources/persistence-app.properties
@@ -1,15 +1,13 @@
-# jdbc.X
-jdbc.driverClassName=org.hsqldb.jdbcDriver
-#jdbc.url=jdbc:hsqldb:file:target/hsqldb/productDb
-#in memory only
-jdbc.url=jdbc:hsqldb:mem:testdb;DB_CLOSE_DELAY=-1
-jdbc.user=sa
-jdbc.pass=
+x# JDBC Properties
+jdbc.driverClassName=com.mysql.cj.jdbc.Driver
+jdbc.url=jdbc:mysql://devops.mysql.database.azure.com/db
+jdbc.user=hamza
+jdbc.pass=admin123_
 
-# hibernate.X
-hibernate.dialect=org.hibernate.dialect.HSQLDialect
+# Hibernate Properties
+hibernate.dialect=org.hibernate.dialect.MySQLDialect
 hibernate.show_sql=true
-hibernate.hbm2ddl.auto=create-drop
+hibernate.hbm2ddl.auto=update
 hibernate.generateDdl=true
