+++ src/main/webapp/WEB-INF/view/header.jsp
@@ -54,7 +54,7 @@
                                     <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false"> Admin <span class="caret"></span></a>
                                     <ul class="dropdown-menu">
                                         <li><a href="./users">Manage Users</a></li>
-                                        <li><a href="./catalog">Manage Catalogue</a></li>
+                                        <li><a href="./poiList">Manage Map Points</a></li>
                                     </ul>
                                 </li>
                             </c:if>
